refactor: put jetty factory as anonymous class to keep everything related closely together in one file
